Build init with shared libraries
Recovery partition now supports shared libraries. Therefore, init can now be built as a dynamic executable both for normal and recovery modes.

To save save in the recovery mode, not all libs are dynamically linked. Libs that are only used by init in the recovery mode are still statically linked.

Note: init is still a static executable for legacy devices where system-as-root is not on, because the dynamic linker and shared libs which are in /system are not available when init starts.

Bug: 63673171
Test: `adb reboot recovery; adb devices` shows the device ID
Test: device boots to the UI in normal mode.
Test: do that for both walleye (system-as-root) and bullhead (legacy
ramdisk).

Change-Id: I30c84ae6a8e507e507ad0bb71acad281316d9e90